+++ functions/accountFunction.ts
